Заработали все методы. Нет Lazy инициализации

+++ src/main/java/com/example/PP_2_3_1_SpringMVC_Hibernate_Boot/model/Role.java
@@ -13,7 +13,7 @@ public class Role implements GrantedAuthority {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -22,12 +22,12 @@ public class Role implements GrantedAuthority {
     public Role() {
     }
 
-    public Role(int id, String name) {
+    public Role(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Role(int id) {
+    public Role(Long id) {
         this.id = id;
     }
 
@@ -40,11 +40,11 @@ public class Role implements GrantedAuthority {
         return this.name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
